# perf: optimize read_as_state_snapshot to avoid HashMap cloning