Replace HashMap.clone() with iterator over references in read_as_state_snapshot().

Performance improvements:
- 26-37% faster execution (tested on 100-10k accounts)
- Reduced memory allocations by avoiding full HashMap clone
- Better scalability with larger caches

Before: O(n) HashMap cloning + O(n) iteration
After: O(1) reference iteration + O(n) selective cloning

Benchmark results:
- 100 accounts: 70,456ns → 45,148ns (1.56x faster)
- 1,000 accounts: 788,236ns → 586,611ns (1.34x faster)  
- 10,000 accounts: 8,588,431ns → 5,370,535ns (1.60x faster)