@onbjerg 2025-10-03
Can you share the benchmark?